requirements.txt
# Dependências principais
selenium>=4.15.0
webdriver-manager>=4.0.0

# Para o pipeline Airflow
apache-airflow>=2.7.0
pandas>=2.0.0

# Para desenvolvimento e testes
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Para logging estruturado
python-json-logger>=2.0.0

# Para configuração
python-dotenv>=1.0.0
pyyaml>=6.0.0
